Made NavModal it's own separate element. Tweaked some nav structure depending on who's logged in.
