+++ src/Components/Projects/ProjectBuilder/ProjectBuildWindow.js
@@ -12,7 +12,7 @@ import ImageIcon from '@material-ui/icons/Image';
 import EmojiPeopleIcon from '@material-ui/icons/EmojiPeople';
 import screenShotExample from '../../../img/screenShotExample.png';
 import teacherShow from '../../../img/teacherShow.png';
-import {UserContext} from '../../../userContext';
+import { UserContext } from '../../../userContext';
 
 export default function ProjectBuildWindow(props) {
 
@@ -94,9 +94,18 @@ export default function ProjectBuildWindow(props) {
                 </div>
             )
         } else if (props.contentId === 3) {
+
+            let x = 88;
+
             return (
                 <div className="projectContentDiv">
-                    <h2>Build your Project</h2>
+                    <br/>
+                    <iframe 
+                    src="https://scratch.mit.edu/"
+                    width={x*16}
+                    height={x*9}
+                    frameborder="0"
+                    title="Scratch - Imagine, Program, Share" />
                 </div>
             )
         } else if (props.contentId === 4) {
